refactor functionality to new pipeline file
