+++ cishouseholds/pipeline/input_file_stages.py
@@ -8,7 +8,6 @@ from cishouseholds.pipeline.high_level_transformations import transform_survey_r
 from cishouseholds.pipeline.high_level_transformations import transform_survey_responses_version_1_delta
 from cishouseholds.pipeline.high_level_transformations import transform_survey_responses_version_2_delta
 from cishouseholds.pipeline.high_level_transformations import transform_survey_responses_version_digital_delta
-from cishouseholds.pipeline.high_level_transformations import translate_welsh_survey_responses_version_digital
 from cishouseholds.pipeline.mapping import column_name_maps
 from cishouseholds.pipeline.mapping import survey_response_cast_to_double
 from cishouseholds.pipeline.mapping import survey_response_cisd_cast_to_double
@@ -17,6 +16,7 @@ from cishouseholds.pipeline.timestamp_map import cis_digital_datetime_map
 from cishouseholds.pipeline.timestamp_map import survey_responses_v0_datetime_map
 from cishouseholds.pipeline.timestamp_map import survey_responses_v1_datetime_map
 from cishouseholds.pipeline.timestamp_map import survey_responses_v2_datetime_map
+from cishouseholds.pipeline.translate import translate_welsh_survey_responses_version_digital
 from cishouseholds.pipeline.validation_schema import validation_schemas
 
 survey_responses_v2_parameters = {
